Fix #148 (#149)
* Fix #148

* Safer size check & integer size calculations

* Added a few test images

The blank test image was generated with
    convert -size 800x280 -sampling-factor 2x2 xc:white tests/reftest/images/blank_800x280.jpg

This creates a grayscale image with a sampling factor of 2 and a height that is a multiple of 8 but not 16,
so that it triggers #148.

* Add more test images

The images were generated with imagemagick to match variants of bug #148

    convert -size 16x24 -sampling-factor 2x2 -colorspace gray xc: +noise Random grayscale_16x24_sampling2x2.png
    convert grayscale_16x24_sampling2x2.png -quality 100 -sampling-factor 2x2 grayscale_16x24_sampling2x2.jpg

The idea is that even though they are currently treated the same,
images with a width that is an integer number of blocks are not the same
as images for which the height is an integer number of blocks.

If only the height is not an integer number of blocks, more
optimizations can be applied. This will be the subject of another pull
request.